Phase 2: Addressed 11th review. Dynamic mock scores, robustness guards, individual agent tests, and Mermaid diagram.

+++ src/core/models.py
@@ -12,7 +12,7 @@ class Scene(BaseModel):
 
 class Critique(BaseModel):
     persona: str
-    score: int = Field(ge=0, le=10)
+    score: float = Field(ge=0, le=10)
     comment: str
     suggestions: List[str]
